First time check in
This is the 'world save' of Siege Perilous as of July 2024. Only player accounts have been removed. The world is completely spawned based on OSI patch 13.6. The world is decorated, and includes the Siege Perilous Custom Housing Preview Area. Additionally, many Townships were left in this snapshot so that administrators can see what is possible with the township builder tool. (feel free to delete these.)